add Array/View to MemSys
array is a reservation of space with a head so the user can append/claim spots and understand when it is “full”.
all objs in a view are assumed to be initialized. view has no head.
both view and array have random access, size, operate on object type (T) and are accessed by index (not byte) like a regular array/vector. they both have bounds checking in the random accessors (at the moment).