add FSA and FSAMap
FSA (fixed size allocator) and it’s map are initialiazed on MemMan init, based on setup parameters.
plan is to intercept any request and redirect to appropriately sized FSA Blocks (if they exists, can be configured differently for any project), and automatically use them instead of creating a new block. (this part not implemented yet)